move StatisticsDataStore to the api module
